chore(jest-core): standardize files name
