+++ packages/jest-core/src/__tests__/__fixtures__/watchPluginThrows.js
