+++ packages/jest-core/src/__tests__/runJest.test.js
