+++ packages/jest-core/src/__tests__/watch.test.js
@@ -11,8 +11,8 @@ import {JestHook, KEYS} from 'jest-watcher';
 import TestWatcher from '../TestWatcher';
 
 const runJestMock = jest.fn();
-const watchPluginPath = `${__dirname}/__fixtures__/watch_plugin`;
-const watchPlugin2Path = `${__dirname}/__fixtures__/watch_plugin2`;
+const watchPluginPath = `${__dirname}/__fixtures__/watchPlugin`;
+const watchPlugin2Path = `${__dirname}/__fixtures__/watchPlugin2`;
 let results;
 
 jest.mock(
@@ -84,10 +84,10 @@ jest.doMock(
   {virtual: true},
 );
 
-const regularUpdateGlobalConfig = require('../lib/update_global_config')
+const regularUpdateGlobalConfig = require('../lib/updateGlobalConfig')
   .default;
 const updateGlobalConfig = jest.fn(regularUpdateGlobalConfig);
-jest.doMock('../lib/update_global_config', () => updateGlobalConfig);
+jest.doMock('../lib/updateGlobalConfig', () => updateGlobalConfig);
 
 const nextTick = () => new Promise(res => process.nextTick(res));
 
@@ -583,7 +583,7 @@ describe('Watch mode flows', () => {
   });
 
   it('makes watch plugin initialization errors look nice', async () => {
-    const pluginPath = `${__dirname}/__fixtures__/watch_plugin_throws`;
+    const pluginPath = `${__dirname}/__fixtures__/watchPluginThrows`;
 
     await expect(
       watch(
