+++ packages/jest-core/src/__tests__/watchFileChanges.test.ts
