+++ packages/jest-core/src/__tests__/watchFilenamePatternMode.test.js
