+++ packages/jest-core/src/__tests__/watchTestNamePatternMode.test.js
