+++ packages/jest-core/src/cli/index.ts
@@ -17,15 +17,15 @@ import chalk = require('chalk');
 import rimraf = require('rimraf');
 import exit = require('exit');
 import type {Filter} from '../types';
-import createContext from '../lib/create_context';
+import createContext from '../lib/createContext';
 import getChangedFilesPromise from '../getChangedFilesPromise';
 import {formatHandleErrors} from '../collectHandles';
-import handleDeprecationWarnings from '../lib/handle_deprecation_warnings';
+import handleDeprecationWarnings from '../lib/handleDeprecationWarnings';
 import runJest from '../runJest';
 import TestWatcher from '../TestWatcher';
 import watch from '../watch';
 import pluralize from '../pluralize';
-import logDebugMessages from '../lib/log_debug_messages';
+import logDebugMessages from '../lib/logDebugMessages';
 import getConfigsOfProjectsToRun from '../getConfigsOfProjectsToRun';
 import getProjectNamesMissingWarning from '../getProjectNamesMissingWarning';
 import getSelectProjectsMessage from '../getSelectProjectsMessage';
