+++ packages/jest-core/src/lib/handleDeprecationWarnings.ts
