+++ packages/jest-core/src/lib/isValidPath.ts
