+++ packages/jest-core/src/lib/updateGlobalConfig.ts
