+++ packages/jest-core/src/lib/watchPluginsHelpers.ts
