+++ packages/jest-core/src/plugins/Quit.ts
